Luis's code so far
Hi guys this is what I've done so far. In this commit, I modify the mainActivity a bit take a look at it if there is anything useful you can implement it. I also created the structure for the main settings so far to change the background color and trophy also difficult. Also, I created a Character class that represents the player and the enemy if any of this is helpful let me know.