Luis's code so far
Guys this is what I coded so far. Please check it out and use anything if it is useful. I created a Character class that represents the enemy and player, I also created the base for the main settings. You can change the background color and difficulty so far. Let me know If you have any questions.